# server/openapi.yaml
openapi: 3.0.3
info:
  title: TinyIce API
  description: |
    REST API for TinyIce audio streaming server.

    ## Authentication
    Two authentication methods are supported:

    **Session cookie** — Log in via `POST /login` to get a `sid` session cookie.

    **Bearer token** — Create an API token in the admin panel (`/admin/tokens`) and pass it
    as `Authorization: Bearer ti_...` header. Tokens can have optional expiry dates and
    track last-used time/IP.

    No CSRF tokens are needed for JSON API requests — the `Content-Type: application/json`
    header provides implicit CSRF protection since it triggers a CORS preflight for
    cross-origin requests.

    ## Real-time Updates
    Use the SSE endpoint `GET /events` for real-time metadata, stream status, and
    listener count updates without polling.
  version: "2.0"
  license:
    name: MIT
servers:
  - url: /
    description: This server
...
paths:
...
  /login:
    post:
      tags: [Auth]
      summary: Log in
      description: Authenticate with username and password. Returns a session cookie.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful, session cookie set
        "401":
          description: Invalid credentials
...
  /api/passkey/register/begin:
    post:
      tags: [Auth]
      summary: Begin passkey registration
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: WebAuthn registration options
  /api/passkey/register/finish:
    post:
      tags: [Auth]
      summary: Complete passkey registration
      security:
        - session: []
        - bearer: []
      parameters:
        - name: challengeKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Passkey registered
...
  /api/passkey/login/finish:
    post:
      tags: [Auth]
      summary: Complete passkey login
      parameters:
        - name: challengeKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Login successful, session cookie set
...
  /api/tokens:
    get:
      tags: [Auth]
      summary: List API tokens
      description: Superadmins see all tokens; other users see only their own. Never returns raw token values.
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of token metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    username:
                      type: string
                    created_at:
                      type: string
                      format: date-time
                    last_used_at:
                      type: string
                      format: date-time
                      description: Empty if never used
                    last_used_ip:
                      type: string
                    expires_at:
                      type: string
                      format: date-time
                      description: Empty if no expiry
                    prefix:
                      type: string
                      description: "Truncated token prefix for identification (e.g. ti_a1b2...)"
    post:
      tags: [Auth]
      summary: Create API token
      description: |
        Creates a new API token. The raw token value is returned ONCE in the response
        and cannot be retrieved later — store it securely.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Human-readable label for the token
                  example: CI/CD Pipeline
                expires_at:
                  type: string
                  format: date-time
                  description: Optional expiry (RFC3339). Omit or empty for no expiry.
      responses:
        "200":
          description: Token created — raw token included (shown only once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  token:
                    type: string
                    description: "Raw token (ti_...) — save this, it won't be shown again"
                  name:
                    type: string
    delete:
      tags: [Auth]
      summary: Revoke API token
      description: Only the token owner or a superadmin can delete a token.
      security:
        - session: []
        - bearer: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token revoked
        "404":
          description: Token not found

  # ── Streams ───────────────────────────────────────────────
  /api/streams:
    get:
      tags: [Streams]
      summary: List all streams
      description: |
        Returns all stream mounts the authenticated user has access to.
        Includes both active streams (with a connected source) and configured-but-offline mounts.
        Offline mounts have `live: false` and empty `source_ip`.
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of stream info objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stream"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Streams]
      summary: Create a stream mount
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount, password]
              properties:
                mount:
                  type: string
                  description: Mount path (will be prefixed with / if missing)
                  example: /live
                password:
                  type: string
                  description: Source password for this mount
                  example: secret123
                burst_size:
                  type: integer
                  description: Initial burst size in bytes for this mount
                  example: 65536
                max_listeners:
                  type: integer
                  description: Mount-specific listener cap (0 = use global cap)
                  example: 250
      responses:
        "200":
          description: Mount created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Mount already taken
    put:
      tags: [Streams]
      summary: Update stream mount settings
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount]
              properties:
                mount:
                  type: string
                  description: Existing mount path
                  example: /live
                burst_size:
                  type: integer
                  description: Burst size in bytes for this mount
                  example: 131072
                max_listeners:
                  type: integer
                  description: Mount-specific listener cap (0 = use global cap)
                  example: 500
      responses:
        "200":
          description: Mount settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Streams]
      summary: Delete a stream mount
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mount deleted
  /api/streams/kick:
    post:
      tags: [Streams]
      summary: Kick source or listeners
      description: Disconnect the source client or all listeners from a mount.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount]
              properties:
                mount:
                  type: string
                type:
                  type: string
                  enum: [source, listeners]
                  default: source
                  description: What to kick — the source encoder or all listeners
      responses:
        "200":
          description: OK
  /api/streams/diagnostics:
    get:
      tags: [Streams]
      summary: List recent persisted diagnostics for a mount
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        "200":
          description: Recent diagnostic entries for the mount
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagnosticEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/listeners:
    get:
      tags: [Listeners]
      summary: List active playback listeners
      description: Returns active HTTP and WebRTC playback listeners visible to the authenticated user.
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of active listener snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Listener"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/listeners/disconnect:
    post:
      tags: [Listeners]
      summary: Disconnect a playback listener
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  description: Listener ID to disconnect
      responses:
        "200":
          description: Listener disconnect queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Listener not found
  /api/listeners/move:
    post:
      tags: [Listeners]
      summary: Move an HTTP playback listener to another mount
      description: WebRTC listeners return `disconnect_required` because live rebinding is not implemented for that transport.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, target_mount]
              properties:
                id:
                  type: string
                  description: Listener ID to move
                target_mount:
                  type: string
                  description: Active target mount path
                  example: /backup
      responses:
        "200":
          description: Listener move queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Listener or target mount not found
        "409":
          description: Listener cannot be moved live (for example WebRTC playback)

  # ── AutoDJ ────────────────────────────────────────────────
  /api/autodj:
    get:
      tags: [AutoDJ]
      summary: List AutoDJ instances
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of AutoDJ instances with playback state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoDJ"
    post:
      tags: [AutoDJ]
      summary: Create AutoDJ instance
      description: Creates a new AutoDJ, starts it, scans the music directory, and begins playback.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, mount, music_dir]
              properties:
                name:
                  type: string
                  example: My Station
                mount:
                  type: string
                  example: /radio
                music_dir:
                  type: string
                  description: Absolute path to music directory on server
                  example: /home/music
                format:
                  type: string
                  enum: [mp3, opus]
                  default: mp3
                  description: Output format. Use `opus` for Ogg/Opus streams.
                bitrate:
                  type: integer
                  default: 128
                loop:
                  type: boolean
                  default: true
                inject_metadata:
                  type: boolean
                  default: true
                visible:
                  type: boolean
                  default: false
                  description: Show on public explore/landing pages
      responses:
        "200":
          description: AutoDJ created and started
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      tags: [AutoDJ]
      summary: Delete AutoDJ instance
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AutoDJ stopped and deleted
        "404":
          description: AutoDJ not found
  /api/autodj/play:
    post:
      tags: [AutoDJ]
      summary: Start / resume playback
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playing
        "404":
          description: Streamer not found
  /api/autodj/pause:
    post:
      tags: [AutoDJ]
      summary: Pause playback
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
  /api/autodj/next:
    post:
      tags: [AutoDJ]
      summary: Skip to next track
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skipped
  /api/autodj/shuffle:
    post:
      tags: [AutoDJ]
      summary: Toggle shuffle mode
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shuffle toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  shuffle:
                    type: boolean
  /api/autodj/loop:
    post:
      tags: [AutoDJ]
      summary: Toggle loop mode
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Loop toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  loop:
                    type: boolean

  # ── Playlist ──────────────────────────────────────────────
  /api/autodj/playlist:
    get:
      tags: [Playlist]
      summary: Get playlist
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playlist items with metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistItem"
    post:
      tags: [Playlist]
      summary: Add files to playlist
      description: Add one or more files. Directories are recursively scanned for audio files.
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mount:
                  type: string
                path:
                  type: string
                  description: Single file/directory path (relative to music dir)
                paths:
                  type: array
                  items:
                    type: string
                  description: Multiple file paths
      responses:
        "200":
          description: Files added and playlist saved
    delete:
      tags: [Playlist]
      summary: Remove track from playlist
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          schema:
            type: string
        - name: id
          in: query
          schema:
            type: integer
            description: Track index in playlist
      responses:
        "200":
          description: Track removed
  /api/autodj/playlist/clear:
    post:
      tags: [Playlist]
      summary: Clear entire playlist
      security:
        - session: []
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mount:
                  type: string
      responses:
        "200":
          description: Playlist cleared
  /api/autodj/playlist/reorder:
    post:
      tags: [Playlist]
      summary: Reorder playlist item
      security:
        - session: []
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mount, from, to]
              properties:
                mount:
                  type: string
                from:
                  type: integer
                  description: Source index
                to:
                  type: integer
                  description: Destination index
      responses:
        "200":
          description: Item reordered

  # ── Queue ─────────────────────────────────────────────────
...
  /api/autodj/files:
    get:
      tags: [Library]
      summary: Browse music directory
      description: Lists audio files and subdirectories. Also shows saved .pls playlists at root level.
      security:
        - session: []
        - bearer: []
      parameters:
        - name: mount
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
            description: Subdirectory path relative to music dir (empty for root)
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileEntry"

  # ── Relays ────────────────────────────────────────────────
...
  /api/tenants:
    get:
      tags: [Tenants]
      summary: List tenants
      security:
        - session: []
        - bearer: []
      responses:
        "200":
          description: Array of tenant objects
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
      description: Session cookie obtained via POST /login
    bearer:
      type: http
      scheme: bearer
      description: API token created in the admin panel (format ti_...)
  responses:
    Unauthorized:
      description: Not authenticated — session cookie missing or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: created
        mount:
          type: string
...